requirements.txt
black
click
pytest
pyYAML
tox
wheel
